+++ docs/news.md
@@ -11,11 +11,12 @@ Released 2024-04-??
 
 - `jsonrpcserver` dep is now optional, enabling Python >= 3.12
     - Add Python 3.12 job to CI
-- new useful commands:
+- new useful commands to normalize/recognize drawings:
     - `texnomagic drawing recognize`
     - `texnomagic drawing normalize`
     - `texnomagic symbol normalize`
     - `texnomagic abc normalize`
+- new `texnomagic paths` command to show/create/open alphabets paths
 - new Drawing, Symbol, and Alphabet reference docs using `mkdocstrings`
     - add docstrings and type hints - nicer code AND docs
     - bump minimal Python to 3.10 due to typing improvements